[Refactoring] Moving IModeManagedInteractor implementation from FuzzyGazeInteractor to GazeInteractor
